fix: arranque prod del multi-tenant (YAML security duplicado + key_version)
Dos defectos del commit 0080795 que abortaban CUALQUIER arranque con perfil
prod (incluido el modo GLOBAL del ambiente de certificacion); solo pasaban
desapercibidos porque los tests unit no bootean el contexto prod y el
contenedor de cert seguia con el jar pre-commit:

- application.yml tenia DOS claves `app.security:` (rate-limit y master-key)
  -> snakeyaml lanza DuplicateKeyException. Fusionadas en un solo bloque.
- V13 definia `key_version` como SMALLINT pero la entidad CertificadoEmpresa
  lo mapea como `int` -> Hibernate schema-validation (ddl-auto=validate en
  prod) aborta con "wrong column type". Migracion a INTEGER.

Verificado en runtime: arranque limpio en GLOBAL y POR_EMPRESA, y dry-run
POR_EMPRESA contra maullin ACEPTADO (factura 33, TrackID 0253303236).

+++ backend/src/main/resources/db/migration/V13__certificado_y_resolucion_por_empresa.sql
@@ -26,7 +26,7 @@ CREATE TABLE certificado_empresa (
     valido_desde     TIMESTAMPTZ  NOT NULL,
     valido_hasta     TIMESTAMPTZ  NOT NULL,
     huella_sha256    VARCHAR(64)  NOT NULL,
-    key_version      SMALLINT     NOT NULL DEFAULT 1,
+    key_version      INTEGER      NOT NULL DEFAULT 1,
     activo           BOOLEAN      NOT NULL DEFAULT TRUE,
     creado_por       VARCHAR(180),
     creado_en        TIMESTAMPTZ  NOT NULL,
